Generalize disk utils as storage Utils for use in browser
The loadJson and loadFile helper functions now accept cross-platform
DataStorage implementations. They should now package for use in
in browser contexts. By moving them into their own file, this is now
achievable since the 'fs' node imports can stay in the disk.js file,
which still contains node-specific helpers.

test plan:
Unit tests have been migrated and flow typechecks successfully.
